added mulitple barcodes per ticket, reworked a bit a logic, added some comments for clarity
Added business rule:  Tickets can contain multiple barcodes
now a tickets have array of barcodes that are checked for listing/buying
updated the logic a bit on in setListingForSale() and buyTicket(()

Thank you for the opportunity! I really enjoyed this assignment over the course of the day and I look forward to a review.